# sample_apis/shop_openapi.json
{"openapi":"3.1.0","info":{"title":"Fake Store API","description":"This API simulates a fake store experience. It offers endpoints to list available store items and to purchase a specific item. The detailed metadata is tailored for LLM agentic orchestration through semantic Swagger docs.","version":"1.0.0"},"paths":{"/store/items":{"get":{"summary":"List Store Items","description":"Retrieve a list of all available items in the fake store. Each item includes its unique ID, name, description, price, and applicable tax.","operationId":"list_store_items_store_items_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/StoreItem"},"type":"array","title":"Response List Store Items Store Items Get"}}}}}}},"/store/buy":{"post":{"summary":"Purchase an Item","description":"Place a purchase order for a specific store item. Provide the product ID, buyer name, and desired quantity. On successful processing, the endpoint returns an order confirmation with detailed purchase information.","operationId":"purchase_item_store_buy_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PurchaseOrder"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"PurchaseOrder":{"properties":{"product_id":{"type":"integer","title":"Product ID","description":"The unique identifier of the product to purchase."},"buyer_name":{"type":"string","title":"Buyer Name","description":"The name of the customer placing the order."},"quantity":{"type":"integer","title":"Quantity","description":"The number of units to purchase. Defaults to 1.","default":1}},"type":"object","required":["product_id","buyer_name"],"title":"PurchaseOrder"},"StoreItem":{"properties":{"id":{"type":"integer","title":"Product ID","description":"A unique identifier for the store product."},"name":{"type":"string","title":"Product Name","description":"The name of the product."},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Product Description","description":"Detailed description of the product, including its features."},"price":{"type":"number","title":"Product Price","description":"The selling price of the product."},"tax":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Applicable Tax","description":"Optional tax applied to the product's price."}},"type":"object","required":["id","name","price"],"title":"StoreItem"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}},"servers":[{"url":"http://127.0.0.1:8000","description":"Local development server"}]}